feature-BJS2-15701
fix conflicts
