--- src/main/java/faang/school/notificationservice/config/RedisConfig.java
@@ -1,62 +1,0 @@
-package faang.school.notificationservice.config;
-
-import faang.school.notificationservice.listener.CommentEventListener;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-@RequiredArgsConstructor
-public class RedisConfig {
-    @Value("${spring.data.redis.host}")
-    private String host;
-    @Value("${spring.data.redis.port}")
-    private int port;
-    @Value("${spring.data.redis.channel.comment}")
-    private String commentEventNameTopic;
-
-    private final CommentEventListener commentEventListener;
-
-    @Bean
-    public JedisConnectionFactory redisConnectionFactory() {
-        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(host, port);
-        return new JedisConnectionFactory(config);
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
-        RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
-        redisTemplate.setConnectionFactory(redisConnectionFactory);
-        redisTemplate.setKeySerializer(new StringRedisSerializer());
-        redisTemplate.setValueSerializer(new StringRedisSerializer());
-        return redisTemplate;
-    }
-
-    @Bean
-    public ChannelTopic commentEventTopic() {
-        return new ChannelTopic(commentEventNameTopic);
-    }
-
-    @Bean
-    public MessageListenerAdapter commentEventListenerAdapter() {
-        return new MessageListenerAdapter(commentEventListener);
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisContainer() {
-        var container = new RedisMessageListenerContainer();
-
-        container.setConnectionFactory(redisConnectionFactory());
-        container.addMessageListener(commentEventListenerAdapter(), commentEventTopic());
-        return container;
-    }
-}
