--- src/main/java/faang/school/notificationservice/listener/AbstractListener.java
@@ -1,54 +1,0 @@
-package faang.school.notificationservice.listener;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.notificationservice.client.UserServiceClient;
-import faang.school.notificationservice.dto.UserDto;
-import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.connection.Message;
-
-import java.io.IOException;
-import java.util.List;
-import java.util.Locale;
-import java.util.function.Consumer;
-
-@Slf4j
-@RequiredArgsConstructor
-public class AbstractListener<T> {
-    private final ObjectMapper objectMapper;
-    private final UserServiceClient userServiceClient;
-    private final List<NotificationService> notificationServices;
-    private final List<MessageBuilder<T>> messageBuilders;
-
-    protected void handleEvent(Message message, Class<T> type, Consumer<T> consumer) {
-        try {
-            T event = objectMapper.readValue(message.getBody(), type);
-            consumer.accept(event);
-        } catch (IOException e) {
-            String msg = "Exception occurred while parsing event message";
-            log.error(msg, e);
-            throw new RuntimeException(e);
-        }
-    }
-
-    protected String getMessage(T event, Locale userLocale) {
-        return messageBuilders.stream()
-                .filter(messageBuilder -> messageBuilder.getInstance() == event.getClass())
-                .findFirst()
-                .map(messageBuilder -> messageBuilder.buildMessage(event, userLocale))
-                .orElseThrow(() -> new IllegalStateException(String.format("No such event type:%s in message builder.",
-                        event.getClass().getName())));
-    }
-
-    protected void sendNotification(long userId, String message) {
-        UserDto user = userServiceClient.getUser(userId);
-        notificationServices.stream()
-                .filter(service -> service.getPreferredContact() == user.getPreference())
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("No such preference contact"))
-                .send(user, message);
-    }
-
-}
