+++ src/main/java/faang/school/notificationservice/listener/CommentEventListener.java
@@ -3,25 +3,32 @@ package faang.school.notificationservice.listener;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.dto.CommentEvent;
+import faang.school.notificationservice.dto.UserDto;
 import faang.school.notificationservice.messaging.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Locale;
 
-public class CommentEventListener extends AbstractListener<CommentEvent> implements MessageListener {
+@Component
+public class CommentEventListener extends AbstractEventListener<CommentEvent> implements MessageListener {
 
-    public CommentEventListener(ObjectMapper objectMapper, UserServiceClient userServiceClient, List<NotificationService> notificationServices, List<MessageBuilder<CommentEvent>> messageBuilders) {
-        super(objectMapper, userServiceClient, notificationServices, messageBuilders);
+    public CommentEventListener(ObjectMapper objectMapper,
+                                UserServiceClient userServiceClient,
+                                List<NotificationService> notificationServices,
+                                List<MessageBuilder<CommentEvent>> messageBuilders) {
+        super(objectMapper, userServiceClient, messageBuilders, notificationServices);
     }
 
     @Override
     public void onMessage(Message message, byte[] pattern) {
         handleEvent(message, CommentEvent.class, event -> {
             String text = getMessage(event, Locale.getDefault());
-            sendNotification(event.getPostAuthorId(), text);
+            UserDto userDto = userServiceClient.getUser(event.getPostAuthorId());
+            sendNotification(userDto, text);
         });
     }
 }
